pyproject.toml
[project]
name = "cryptocurrency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.5",
    "defusedxml>=0.7.1",
    "lxml>=6.0.4",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "sortedcontainers>=2.4.0",
    "websockets>=16.0",
]
